requirements.txt
django==3.0.8
psycopg2-binary==2.8.5
djangorestframework==3.10.3